service worker: Add tests for inteception of workers after redirects.
This tests behavior discussed here:
https://github.com/w3c/ServiceWorker/issues/1289

Namely it tests when a request for a worker goes through a redirect
chain:
1) On redirect from A -> B, whether the service worker at B
sees the request.
2) After the final redirect, which service worker controls the
resulting client.

The tests are written as specified today. Therefore, Firefox
passes this test (verified in Nightly) and Chrome does not.

Currently it only tests shared worker but dedicated worker can
be added in a follow-up patch.

Bug: 731604, 724371
Change-Id: Id3b1ea8b952760be0ef9917f2c6a3afe60ca1fb5